+++ src/main/java/kitchenpos/menu/model/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.model;
 
 import java.math.BigDecimal;
 import java.util.Objects;
