+++ src/main/java/kitchenpos/menu/repository/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.model.Menu;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
     @Query("SELECT COUNT(m) FROM Menu m WHERE m.id IN :ids")
