+++ src/main/java/kitchenpos/menuproduct/repository/MenuProductRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.repository;
+package kitchenpos.menuproduct.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menuproduct.model.MenuProduct;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
     List<MenuProduct> findAllByMenuId(Long menuId);
