+++ src/test/java/kitchenpos/menu/domain/MenuVerifierTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -7,6 +7,10 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import kitchenpos.domain.Product;
+import kitchenpos.menuproduct.model.MenuProduct;
+import kitchenpos.menu.model.MenuVerifier;
 
 class MenuVerifierTest {
     @DisplayName("메뉴 등록 시, 메뉴에 속한 상품 금액의 합은 메뉴의 가격보다 크거나 같아야 한다.")
